+++ libs/flexIPC/fleXdEpoll/ChangeLog.md
@@ -2,6 +2,20 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+## [0.1.0] - 2018-07-17
+### Added
+- New setOn... functions for FleXdEvent and FleXdTimer [Martin Strenger]
+- Added FleXdEvent::uninit() function [Martin Strenger]
+
+### Changed
+- Fixed missing file descriptor closing in FleXdTimer::stop() function [Martin Strenger]
+- Fixed behaviour for multiple calling of FleXdEvent::init() function [Martin Strenger]
+- Fixed behaviour for multiple calling of FleXdTimer::start() and FleXdTimer::stop() functions [Martin Strenger]
+- Fixed epoll_wait deadlock [Martin Strenger]
+
+### Removed
+- None
+
 ## [0.1.0] - 2018-06-12
 ### Added
 - New functionalities FleXdTimer, FleXdEvent, FleXdTermEvent, FleXdSignalHandler [Martin Strenger]